requirements.txt
hydra-core==1.3.2
matplotlib==3.7.2
numpy==1.25.2
omegaconf==2.3.0
Pillow==10.0.0
Pillow==10.2.0
pydantic==2.6.3
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.66.1
